+++ src/exception/focus-exception.js
@@ -1,15 +1,15 @@
-let CustomException = require('./CustomException');
+const CustomException = require('./custom-exception');
 /**
- * Class standing for the FocusException exceptions.
- */
+* Class standing for the FocusException exceptions.
+*/
 class FocusException extends CustomException {
 	/**
-	 * Exception constructor..
-	 * @param messgae {string} - Exception message.
-	 * @param options {object} - Object to add to the exception.
-	 */
+	* Exception constructor..
+	* @param messgae {string} - Exception message.
+	* @param options {object} - Object to add to the exception.
+	*/
 	constructor(message, options) {
-		super("FocusException", message, options);
+		super('FocusException', message, options);
 	}
 }
